+++ appwebFAP/src/main/java/freewill/nextgen/data/JamShowEntity.java
@@ -43,7 +43,7 @@ public class JamShowEntity implements Serializable, Cloneable {
     private float PVTotal=0;
     private float puntuacionTotal=0;
 	private int clasificacionFinal=0; //Puesto en el que quedó el patinador en resultado final
-	private float sumaPonderada = 0;
+	private double sumaPonderada = 0;
 	
     /**
      * Default constructor. 
@@ -319,11 +319,11 @@ public class JamShowEntity implements Serializable, Cloneable {
 		this.puntuacionTotal = puntuacionTotal;
 	}
 
-	public float getSumaPonderada() {
+	public double getSumaPonderada() {
 		return sumaPonderada;
 	}
 
-	public void setSumaPonderada(float sumaPonderada) {
+	public void setSumaPonderada(double sumaPonderada) {
 		this.sumaPonderada = sumaPonderada;
 	}
